Add a conditional around the M900 command
The intent of the conditional in the filament profiles is to allow a PrusaSlicer user to use the filament profiles with printers that do not use Linear Advance 1.0.  With machines based on LA 1.5 (Marlin 1.1.9) or running RepRapFirmware the values passed to M900 by the default profiles would be far too high.